Only create a Tray icon on linux/windows
Currently, the code creates a tray on all platforms. On MacOS X
this creates an "application menu" in the upper right. However, this
is not needed at this time. Additionally, it requires a png to be
bundled which is currently not happening on MacOS X. This causes the
application not to launch at all.

This commit disables the creation of a tray in order to allow the app
to launch. In the future, we may add this functionality back.