Ensure we only persist state on important changes
Currently, we are persisting the entire state every time a key is
pressed. This is leading to a significant slowdown while typing
if there are any messages at all to be persisted. This commit adds
a filter to the vuex-persist plugin to only persist when certain
mutations are called.

This is a first pass at what mutations should trigger a persist. Future
work should look more closely at this and see if we cannot attach
the vuex-persist to only relevant stores.