Sort chatlist by how much "unread" money a contact has sent
Currently, the chatlist is sorted by order that you that contacts
were added via either sending or receiving messages. However, the
point of using this chat client is to be able to prioritize senders
who value your time.

This commit changes the sort order to be based on total unread value
of the messages you've been sent.