Fix chatlist sorting to sort read chats by last read time
Previously, there was no tie-breaker for chats that were read. However,
it did pull the selected chat to the top of the list. This caused the
UI to be a little jumpy. To fix this, we sort any chats that have
the same values by the last time their chat was read.