Format unread value to have units and add back total unread messages
Currently, we were only displaying the total number of unread sats.
This was somewhat unwiedly for large numbers, and it wasn't clear
to users what the number was intended to be. This commit adds the
appropriate currency formatting, and also adds back another chip which
shows the total number of unread messages.