Fix generation of change outputs so that fees are correct
Occasionally, we would see a transaction fail to relay because the
fee per byte was too low. This was because we were not considering
the size of the final change output in our calculations for generating
extra stamp-sized UTXO outputs

This commit updates the calculates to ensure that extra UTXO is
considered when calculating how many excess UTXOs to create.