Update stealth txn processing for new proto specs (#24)
In an effort to obfuscate transaction amounts, we have added bundling
to the stealth entries. This allows transfered amounts to be split
over multiple bitcoin cash transactions.

This commit adds support for the new protobuf specification that allows
that. It supports on the receiving end arbitrary splits, but on the
sending end only does one output and one transaction currently.