Make chat area not re-render when switching tabs
Herin we render multiple chat components for each contact, but only
show the active one. This ensures these chats aren't re-rendered when
switching contacts. This is at the expense of memory and initial load
times. However, much of this can be obviated with more adjustments to
how vuex is being used.