Simplify and optimize message timestmap handling
This commit removes the global timer which currently goes through
the vuex store. This causes the need for extra indirection and storing
of the data to disk when it is not needed. This commit instead
uses a local timer, and switches to using momentjs instead of our
own formatting library.