Remove use of tabHeight to fix heights in layout
This commit removes the need to pass around and use tabHeight within
the components of the layout by ensuring the outer containers are all
set to max height. Additionally, the q-page component is set to the
maximum size view a `style-fn` which receives the viewport header/footer
size as an offset.