Fix reset of the chatarea height when not shown
When the time is updated, the height of the chatboxes are also
reactively updated. However, since the contents are not shown the
chat input has a height of zero. This causes the chat message box
to take 100% of the viewport. When switching back to hidden viewports
it takes awhile for it to update again. This looks bad.

As a temporary fix, we set the minimum value to 50 (which is the same)
as the chat input on most devices.