Fix bug with adding a new contact (#64)
We incorrectly were setting the wrong key for `messages` during a
previous refactor. Additionally, the contact wasn't being created when
trying to "send local" for the first time to a new person. These two
issues made it so you could not talk to new people. This commit address-
es both of those issues by fixing the key and adding a check to create
the contact. It also adds a migration to fix the data store of people
who downloaded the last release.