Add a wallet module and create a wallet class (#87)
This commit adds a wallet class that is independent of the Vuex code.
The goal of this is to work towards making the CashWeb portions of the
code independent of Vuex. Once this is done, it will be possible to
export these as an npm module, and an SDK for writing other software.