Migrate to using wallet class (#92)
This commit removes the unnecessary wallet functionality from the wallet
Vuex store, and switches to using the wallet class which was introduced
in a previous commit. The long-term goal here is to get all the CashWeb
protocol interaction detangled from the Vue/Vuex code so that it can be
used in headless applications.