Integrate new relay client with the rest of the codebase
This commit integrates the factored-out relay client into the rest of
the codebase and Vuex store for state. The goal here is to create self-
contained libraries which could be used to build bots and other items
things in the future. Additionally, these components will be able to be
tested in the future more easily, and more clear what state they
depend on when refactoring.