Remove spend utxos while syncing (#100)
Currently, we add all UTXOs we've received while syncing, and then
run through them vs the electrum server. This requires that we slam
the electrum server asking for histories for ever address we've ever
used. This is not scalable, and is particularly inefficient. Instead, we
can simply remove anything we've recorded as being spent from our UTXO
set.