Major refactor of message construction APIs (#105)
This commit is a major refactor of the message construction APIs that
deduplicates a lot of code for generating payloads and sending messages.
This was motivated by the need to refactor a significant number of
different locations in the code when updating logic for sending
messages to deal with edge cases. As such, there is now one place that
generates and sends message payloads.