Reset state of schema versions are different
This commit causes chat messages and other actions to not cause disk
writes. The only thing that are stored are necessary items to restore
state from the relay server. We need to figure out how to only store
chat messages in bulk once they are loaded. Possibly some timer to
autosave in the future.