Update LICENSE and Github pull_request_template
It's been a frequent practice that various companies are whitebox-
relabling open source projects within the open source ecosystem. This
is a large hinderance to the professional maintenance of software
projects. As such, we are changing from an MIT license, to something
which permits the free distribution, and modification of source code
*only*.

We welcome competition within the Stamp ecosystem, and the backend
services and protocol are free and open. However, we wish to force
companies to participate in *this* project, by relicensing or directly
contributing -- rather than producing poorly maintained forks.

My hope is that this project will able to self-sustain through a
commercial offering. Having this project copied and rebranded by for-
profit companies