Ensure we choose a new electrum server on reconnect
A fair number of electrum servers seem to have bad uptime. This commit
addresses this by trying a new electrum server upon reconnect. Ideally,
we use multiple servers so as to not give away too much information
about which addresses we are using. However, this should work just to
ensure we are able to remain connected to the electrum servers.