Fix sizing issues with input/reply box (#139)
Currently, we're having some issues with how the input boxes respond reactively. This is because when the main areas are not shown, their height is zero.  It takes some triggering to fire off the resize events, resulting in a bad user experience. This commit uses the column flexbox layouts to avoid that issue. Although, it's not clear how the `q-scroll-area` is automatically adjusting its height appropriately. 

Co-authored-by: Shammah Chancellor <shammah.chancellor@gmail.com>