First pass at using containerized layouts for chat area
We were originally calculating height/widths manually through events.
This resulted in slow resize times as it was not using CSS. We then have
switched to using columns/rows using flexbox. However, this method does
not allow for `q-page-sticky` to work properly. This branch is the start
of and attempt to use an inner-layout with the input box and name header
in side that layout. This allows for the stickies to show up in the
appropriate location.