Fix issue with reconnect logic attempting to reconnect to old servers
When we are disconnected from a wallet server, the keep-alive logic
hangs around attempting to reconnect to the old server. Even though,
we have likely already connected to a different server. This commit
both removes the prompt-to-reconnect (as it does nothing now), and the
logic to continue to keepAlive on the old server.