Use arrays for chat message handling (#165)
Currently, we're using a hashmap for messages per contact. However,
these should be unordered. It's unclear how these are showing up in the
correct order (beyond the fact that they are added sequentially). In the
future this may change for efficiency purposes, additionally, we want to
be able to omit some parts of `ChatMessage` displays -- as well as add
in separators in some cases. Thus, this commit switches to using a
normal ordered array for each chat's messages.