Basic username coloring in chats (#176)
Motivation:
Suppose a malicious actor copied someones profile and messages you - it would be tough to distinguish the original from the fake purely based on the difference in the address. Especially when one can grind on addresses to achieve similarity.

In order to help users distinguish between the friend and the malicious actor we introduce a colour scheme whereby a colour is selected using H(address || salt).

Implementation:
The current implementation should work as an basis for future iteration. We currently pick the hue and saturation from the first two bytes of the digest and fix the lightness to achieve a colour from HSL. The lightness is fixed to avoid very bright or very dark colors which make hurt accessibility.

The salt is currently hardcoded into the constants.js but in future should be deterministically generated from the identity private key or randomly generated at setup then stored.