Ensure relay client is allowing JavaScript eventloop to tick
Currently, the relay client refresh locks the client up. This locking
happens because the event loop is only allowed to process an event on
io, or by explicit yielding. This commit ensures the event loop is
yielded to every 5 messages that are processed. This allows the GUI
to keep responding to user input, and appear to have loaded immediately.