Ensure wallet init happens on load (#188)
This commit ensure wallet init happens on load.  However, it also moves
the UTXO refresh call into the relay client. This ensures that the time
consuming update occurs only when it would be relevant. That is, after
all messages have been loaded and UTXOs that were sent out being
removed.