Do not refresh UTXO state for stamps/stealth txns
Due to a previous update, the client now properly updates its internal
UTXO set state from the relay server. As such, spamming electrum with
significant numbers of `listunspent` requests is no longer necessary.
It is still possible that due to reorgs that we get out of sync.
However, this can be solved by fixing those utxos and re-trying when
we attempt to use them to send a message.