Ensure notifications aren't sounding for already seen messages (#195)
Currently, we are reloading all the state from the relay server.
We are also not showing read messages as "unread" now that we are
storing `lastRead` time properly. However, the desktop notifications
are still firing. This commit fixes that by checking last read for a
given contact when looking to send a desktop notification.