Fix `Cannot read property lastRead of undefined` (#198)
When first receiving a message from a new contact, the chat is undefin-
ed. As such, the `chats[addr]` is undefined, and it impossible to look
up the `lastRead` field. This commit checks to see if the "chat" exists
and returns zero if not. That prevents this error from being hit.