Fix default chat message list to be appropriate type (#199)
In a previous refactor the messages store (per chat) was switched
to an array from an object. However, a few places were missed because
javascript is not typesafe. This commit addresses the problem when a new
contact is added. In the future this should be address more holistically
by moving these files to typescript.