Fix exit behavior on mac (#200)
In order to stop the app from exiting, we were preventing the default
handler for the `close` event in electron. However, on a mac this makes
it so that the app won't exit on shutdown or a menu shutdown. This
commit updates the main event handlers to check if the app is being
`quit` on MacOS (vs just having the window closed).