Fix reload (#205)
We previously overloaded hyperlink navigation - instead of opening the new link in Electron, it opens it in an external browser. This was required as we don't want hyperlinks in text chat to be opened in Electron.

The unintended consequence of was that the navigation associated with reloading the app would open in the browser (at http://localhost:8080). This fixes this by catching that behavior.