Fix floating button behavior of chats
Due to a previous change, the floating scroll button was broken. This
commit fixes the event handlers to properly update the `bottom` data
value for the chat. Additionally, we are no longer using
`ScrollObserver` as the `ScrollArea` has it's own event which is more
accurate and responsive. It also properly fires on resizing the
scroll area, allowing for adjustments when `v-show` is updated.