Better input box automatic selection (#220)
Auto focus on inputs where needed. Added ability to press enter after filling in a user's address to add them as a contact. I wasn't sure if we would want to do that for the Send to Address menu since that's a little more risky. I'll create another story to allow up and down arrows to navigate better when searching for a contact.