Fix tab rendering for the transaction dialog and add type info (#231)
Currently, the rendering of the tabs in this dialog is incorrect. It
does not have the `q-tab-panel` wrapped with the appropriate control,
and sharing a model with the `q-tabs` element. This commit fixes
that issue, as well as adding a type field and renaming the menu item.

The second two changes are because this dialog also shows other types
of outpoints attached to a message.