Clean up various places where undefined is not checked (#249)
Due to the way messages are sent to the client, they may have various
malformations in them. The client as such, needs to be robust to invalid
messages. This commit adds more checks for undefined values that cause
exceptions in JavaScript. These particular spots have been observed in
practice when for whatever reason, some entries are removed from the
keyserver.