Use level storage wrapper for saving messages (#248)
Previously, we used the Vuex-persistence store for messages. However,
this uses a single localstorage key for storing all data as one
serialized JSON blob. The limitations of this caused undefined behavior
after a certain amount of messages were received. Thus, we are switching
to manual storage against leveldb/indexdb. This wrapper allows us to
implement once, and have support in both node and electron/browsers.

Node support is important as we also want to use these as libraries in
the future to enable people to write bots and other headless wallets.