Address issues with reconnect logic on load (#250)
The electrum-cash `client.connect` method never resolves its promise
when connections fail. Thus the setup prior to MainLayout loading was
never succeeding if the first attempted Electrum server was offline.
This commit removes the `await` and updates downstream logic to properly
handle the need to reconnect to the electrum server immediately at
startup.