Replace electrumClient with a promise to a client
Currently, we set the electrumClient field in a variety of places once
as soon as the client has been created. However, this means that we may
attempt to make requests before connection is completed. This commit
changes the system to provide a promise to a client, which gets resolved
or rejected, when the client has actually connected. This ensures that
the client is ready before making requests, and that init() is
properly called in the wallet every time the electrumClient is restarted