Ensure constructTransaction[Set] does not reuse UTXOs (#280)
Currently, we are not properly awaiting freezeUTXO which would hyptoth-
etically allow a frozen UTXO to be selected again before it froze.
Additionally, constructTransaction was not properly filtering out frozen
UTXOs from construction. This meant that early on in the setup process
(and possibly ather places) you could not generate two of the same
output-size transactions back-to-back before the wallet was updated from
the electrum subscriptions.