Fix keyboard navigation of contacts
A previous commit enabled the contact search dialog to be opened with
Cmd-K. However, there were some oddities in the behavior when pressing
enter, or having accidentally had an address which was not part of the
search results. This commit removes some complexity, and adds the
necessary watchers to make sure that the address is always defined.

Fixes #221, #225, #287