Do not open/close leveldb repeatedly (#300)
Leveldb does various checks and cleaning when it is opened and closed.
This creates unnecessary overhead every time we want to get or set a
key. As such, this slows transaction generation considerably. This
commit removes that process and ensures the db is gracefully closed when
possible.

Part of fixing #298