Use random UTXOs instead of deterministic selection (#312)
Due to deterministic coin selection, sometimes the wallet can get "stuck." This is undesirable during error retries. This commit ensures that a new UTXO will be selected each time constructing a transaction set is retried. This has a beneficial effect on chainalysis as well in some regards. Previously, sometimes revealed the total amount sent if there was a UTXO larger than the current one. This ensures we sometimes do select a random UTXO for small amounts. However, there is more work to be done to ensure transaction set building is more anonymous.