Ensure there are at least two transactions in a transaction set
In order to ensure that amounts are obfuscated, as well as stamps, we
now ensure that there are at least two iterations of the transaction
set builder. Each iteration is given a randomized amount bewteen
`[minimumOutputAmount, amount]`. This ensures that we won't reveal an
amount when there is a UTXO larger than the amount we were intending to
send.