Ensure we attempt to use sufficiently sized UTXOs first (#314)
Because we now split transaction bundles into a set number of fragments,
it is better to now try to use a random UTXO that is bigger than
whatever amount was determined to be used. This ensures we don't try
to create 1-input -> 1-output transactions for no particular reason.
Doing so would waste users funds on fees.