Add support for multiple change outputs back to tx construction
In a previous commit, we removed support for generating more than
1 change transaction per txn. This was done because the old change
generation code made it very obvious which output was being sent to
another person. However, this is undesirable because we
want to extra UTXOs so we do not hit chain limits. This commit adds
this functionality back, but with a mechanism to appropriately obfuscate
the actual output from the change.