Ensure profile is setup with keyservers and relay servers on startup (#326)
Currently, we only ever upload our profile, or keyserver metadata when
going through the setup. This means that if for some reason we change
backends our client will not automatically update keyserver/relay server
data.

Thus, this commit factors out the code from the Setup steps. Then,
we setup the MainLayout created hook to check for existing profile/
keyserver data, and update the backends if it is not there.