Use isomorphic websocket module for connecting to relay server
This will allow connecting from both Electron and from Capacitor, and
web browsers in general. This will greatly improve our ability to
support other platforms. However, we also need to update the relay
server to support the full bearer token specification. The web browser
API for websockets does not allow setting headers, which means the
bearer token cannot be included in the Authorization header when
negotiating a websocket.