More fixes with the use of the newer electrum client (#337)
Some automatic re-subscribing and other functionality was added to the
latest version of the client. This caused some issues with the internal
disconnect handling logic we implemented. We may be able to get rid of
that logic in the future, but for now minor adjustments are made to
fix this particular issue.