Ensure generated seed phrases are always saved, and setup can be resumed (#339)
Due to state errors around how seeds are generated and stored during
setup, the setup could not be resumed if the xpriv had been generated,
but the setup not completed. This commit addresses those issues in order
to ensure that funds are not lost during a partial setup.

However, from the areas touched, it should be apparent that the way in
which this state is being handled is rather tangled. There needs to be
an overarching concept for "setup complete" rather than it being
inferred from various other variables being set.

Fixes #338