Fix subscription handling due to a different change in electrum-cash API
The electrum-cash library was changed to automatically resubscribe to
all script hashes. As part of this, it also throws an exception if there
are any event listeners that aren't registered in the map of listeners
and params. This commit removes the need to have a single listener
on the entire connection object. Instead, we now let the internal
logic for the electrum-cash client handle this.