Ensure UTXOs are cleared from leveldb during Setup (#341)
Not reseting leveldb caused  the balance to not properly refresh due to
the desync between the rehydrated wallet being reset, but not the
leveldb store when the setup regenerates a new wallet.

Thus, we need to clear out the actual wallet at the same time. In the
cuture we should remove explicit UTXO tracking in the Vuex store, and
instead only update the balances as UTXOs are deleted/removed.