Ensure local_modules is copied prior to production build
The electron-builder runs `yarn install --production`, however the
local_modules folder is not copied by default. This means our stub of
bn.js cannot be included and the build errors. This adds the webpack
copy plugin to ensure that it moves the folder into place prior to
the actual build and bundling for electron.