Fix subscriptions around the identity in the sign up flow
We want to allow people's initial paying transaction to also reveal
their idenity. That way, when a friend sends you money to create an
account, they can immediately add you to their contact list. This commit
fixes what was previously an oversight -- We had change the deposit ux
to use this address, but had not updated the subscriptions to function
with this address which was previously different.