Ensure drawer opens on first import (#362)
Currently, the drawer does not open the first time the client is loaded
due to the active chat address not being set. This means on desktop you
can never get to the drawer due to the heading bar not being shown. This
commit introduces a temporary fix of showing the drawer on load. A
better fix would be introducing a news page for people to land on.