Switch codebase to use parameterized network names for display vs API
Currently, Bitcoin ABC is still using the `bitcoincash` address prefix.
However, this is not a long term situation. Additionally, our codebase
hardcodes the network name in many different places. This commit splits
the display of addresses, and the use of the addresses in the API
backend -- and uses one constant everywhere so it'll be easy to swap
out.