Ensure signups for *BRAND NEW* accounts don't fail (#367)
When a person signs up, they need to make two payments immediately.
One for the keyserver, and one for the relay server. However, if they
only have one UTXO, this won't be possible until the change of the first
transaction settles. This commit updates the error handling around the
second payment to retry until there are funds in the wallet due to
the subscriptions from the first transaction being hit.