Add a basic landing page, so you can get to the drawers
Currently, on some small screens there is no way to open the basic
contacts drawer, to be able to do anything from there. This commit
adds a hack to create a home page when the active chat address is not
yet set. It also adds a link so you can get to the home page as needed.