Refactor routes so that various dialogs can be their own pages
Currently, various actions are "Dialogs." However, these do not render
well on mobile devices because the screens are too small. This is a
basic refactor of the Main Layout so that Settings can appear instead
of chats. This may have some performance consequences down the road that
we will need to deal with.