Only replace settings urls when navigating from settings panel
We needed to use route replace, or going back would cause you to end
up with a stack of various settings you'd need to cancel through.
However, when you were on the news route, it would take you all the way
back to the signup screen in some cases. This commit addresses that
problem.