Fix bugs in forwardUTXOsToAddress
Currently, if the UTXOs have already been used, `forwardUTXOsToAddress`
will fail to properly generate and forward a transaction. This commit
fixes the issue by short circuiting. Additionaly, it also scanned every
UTXO looking for the UTXO to delete, rather than doing a simple lookup.
That issue is also fixed in this commit.