Remove the need to recalculate wallet balance from all UTXOs
Currently, we were looping through all UTXOs to calculate the balance
at each visual update. This slows down the frontend quite a bit in
some circumstances. This commit removes that need by keeping the balance
as a cached variable. The rest of the GUI does not depend on any
information about UTXOs directly.