[wallet] Fix transaction building to not just mindlessly retry
The current version does not correct the transaction number in cases
where the transaction failed to build. This commit properly pipes back
the data for the transaction number so the entries in the stamp message
are not invalid. Additionally, the transaction constructor now tries to
add additional inputs rather than retrying mindlessly until failing.