[wallet] ensure transactions are not generating uncessary change outputs
During the last refactor, some of the assumptions around generating
transactions were violated that cased change outputs to always be
produced. This commit remedies that enforcing the appropriate checks
during transaction building. (In particular that we don't require
the amount to be able to generate change always, only when it is more
than double the expected fees.)