[wallet] Ensure UTXOs are never reused in the same transaction
Due to the way splice was being used w/ pickOne, if the `biggerUtxos`
was being used for selecting the UTXO, it would not properly remove it
from the next iteration's possible UTXOs. This could sometimes result in
a UTXO being reused. The resulting transaction would be invalid. This
commit removes the plucking of the UTXO from the array to be always
against the total working set of UTXOs rather than sometimes against
the ephemeral array created by `filter.`