Make change output more accurate and drop limit on minimum output amount
This commit fixes several logic errors in generating change output
amounts as well as drops the limit to the dust limit. This should allow
for a better distribution of change as well as more correct generation.
The last stanza in this commit happened to work to fix calculation
errors in the first loop.