Improve _buildTransactionSetForExplicitAmount input selection
This comit changes a few checks so that the number of inputs added
can result in a transaction with no fees (assuming there are the
appropriate combination of UTXOs). Additionally, it makes it so that
there should often be at least the desired minimum of transactions
unless there are no UTXOs smaller than the amount specified or the
amount specified for construction is less than minimumOutputAmount.
In these cases, one transaction will be built. In the second case,
a cascade of new change outputs will be made to satisfy future
transaction building use cases.