Remove splitting logic outside of _buildTransactionSetForExplicitAmount
Previously, we split the passes for building transactions for explicit
amounts. However, it would be preferrable to build the whole set as
part of the same logic, which allows more transactions to be constructed
without multiple inputs or change outputs. The previous two commits
enable that behavior. This commit removes the logic which was previously
necessary to generate multiple transactions most of the time.