Ensure clicking send and scroll down don't unfocus ChatInput
Currently, the mobile experience for sending messages is fairly terrible
because the click event caused the textbox to become unfocused. This
means the keyboard goes away immediately. This patch changes prevents
the event from proagating and thus changing focus off the input both
for the ChatInput and the scroll down button.

Unfortunately, the behavior is somewhat strange as we now attach
mousedown handlers directly to icons, and lose some styling. I was not
able to prevent the unwanted behavior on q-btn.

Co-authored-by: viet <viet@tribeaccelerator.co>