Update bitcore-lib-xpi type definitions to add more fields/types
The Bitcore lib is extremely flexible in terms of what it lets you pass
around. This update adds a few types that will be needed to convert the
existing wallet code to typescript. The parameter updates and fields
that were added to in fact exist on these types, but are not defined
in the typescript definition files we originally co-opted.