Upgrade to quasar 2, Vuejs 4, and electron 13
This commit is *massive* because there are a lot of breaking changes in
quasar 2 and vuejs 4. Additionally, electron 13 also broke local file
access from the renderer thread. So, we now only use the IndexDB version
of the `level` package even when using electron. However, when the
cashweb suite is used directly from node, it will still use LevelDB
directly.