Fix issue with users unable to scroll up
A pre-emptive `scrollBottom` was added to the scrollHandler in the case
where their might be a race condition between adding messages, and
attempting to scroll. However, this seems to be unnecessary with the new
control flow for `scrollBottom`. Additionally, it was preventing users
from scrolling up using the scroll wheel on their mouse due to events
firing frequently rather than when scrolling stopped.