Clip long words in Chat list captions to prevent overflow
For whatever reason, despite content having a max width, the QScrollArea
still expands. This results in badges being off screen and not visible
to users. This happens only when words that are too long and cannot be
wrapped are being shown. Thus, this commit clips long words that would
be displayed in the caption.