Fix issue where change output was not being generated causing asburd-fee
Occasionally, due to the effort to not generate change outputs which are
the same order of magnitude as the actual amount sent, a change output
would not be generated. This commit ensures that at least one change
output is generated if there is sufficient change for it to be relevant.
Additionally, the output that the additional amount was being randomly
placed into when necessary, was always output `0` and should have been
a random change output. Both of these issues are fixed in this commit.