Refactor UTXO handling on message send to be able to update UTXOs
Currently, we do not check UTXO validity before sending transactions.
This sometimes results in some of the transaction bundle being burned as
some transactions can broadcast correctly; while the final one does not.
This would prevent the relay message from passing through and allowing
the recipient to obtain the message. Now, we pass full UTXO data rather
than just the ID up to the caller. This allows checking to see if they
are unspent without requiring a UTXO lookup. Additionally, this code
also checks the validity of all UTXOs before sending a message.