Convert index file of store modules to typescript
This is both a ts conversion and fixes an issue with importing
typescript store modules. They were not being picked up due to
regex filename magic. We don't have so many different stores that
this magic import setup is particularly worthwhile and it just
adds complexity -- and caused some things to break.