Convert Setup.vue to typescript
Convert Setup.vue to typescript with a few type "holes" with the store
left. These will need to be converted later as the documentation is
unclear how to use the store in a type-safe manner with the VueJS
"options API." It may be necessary to convert to the composition API
in the future.