Convert level-store plugin to typescript (#417)
This commit factores out the level-store plugin for Vuex that is custom
to Stamp. However, it seems the plugin system was not designed to use
RootState in plugins directly to access substate. While this system
makes typescript happy, it seems to not quite be how things are supposed
to work. If you're reading this, and you know how this should function
in order to have a proper reducer please take a stab at it.