Ensure deleted UTXOs are re-added after being deleted
Currently, we download messages potentially out of order. This could
cause the deletion of UTXOs from previous messages to have no effect
if a future message caused the addition of those UTXOs. By keeping an
in-memory cache of what was deleted, and ensuring they are not re-added
we remove order-dependence on message processing.