Fix local message deletion
At some point in the past, the structure of messages in the chats store
was changed to be an array. However, `splice` was still being called
to delete messages from the list of messages stored. This was causing
the `splice` call to hang for an extremely long time as well as causing
messages to not actually be deleted.

This commit addresses that by finding the appropriate index before
slicing, and also by deleting it from the general message index that is
used for displaying and updating message data.