Add AppImage parameters to Quasar.conf.js
In the recent update of quasar there seems to be an issue somewhere such
that electron-builder is not properly passing the icon defaults to
app-builder. As such it was failing with the following error:

```
  ⨯ stat /Users/xxxxx/repos/xxxx/src-electron/icons/icon: no such file or directory
```

Adding the full icon path addresses this issue, as well as a warning
about app categories.