Move UTXO updating to after message refresh
Occasionally, when downloading and synching messages, there are UTXOs
from messages that failed to be deleted when running a remote wipe
of your chat history. However, the outputs are still forwarded to your
HD wallet. Because of this, the current UTXO check code needs to run
after your wallet fully syncs with the remote server.