Factor out message component from Agora.vue page
This commit factors out the agora message component from the main vue
page. This makes it easier to modify independently of the other things
on the page. But, more importantly it is required for recursive
elements. Recursing on message will be required for replies to be able
to function correctly.