Ensure message entries without a payload are ignored (#444)
Currently, if a payload is absent from an entry, protobufs returns an
empty string. This trips an assertion that was added to make typescript
happy and prevents the Agora page from loading if such a message exists.

Somehow someone posted such a blank message to the server. Possibly due
to an unknown bug, or someone is playing with the source.