Refactor the Agora to use router views to dedupe layout
Currently, the Post Page and the main agora page do not share layouts.
This commit refactors the content to share a layout and use a nested
router-view in order to ensure both pages are rendered in the same way.
Additionally, we add v-bind="$attrs" in order to allow events on message
to propagate all the way back up to the router-view and layout.