Fix message filtering in Agora main page (#459)
A previous commit changed the behavior of the Agora fetching routine to
only fetch recent messages. However, this resulted in filtering behavior
not to function correctly as it was processing on the server side. This
commit moves the filtering to the client side and uses v-show instead of
filtering server side so that messages can remain cached locally.