Fix message load times in forum
Due to processing messages while they are part of the state the DOM was
attempting to update at the same time. This caused lots of DOM mutations
and redraws which slows down the loadtimes for forum pages
significantly. This commit mutates the lists outside of the DOM updates,
and then sets the whole message state at once.