Fix p2pkh sends
When converting over to typescript, it turns out that p2pkh sends do not
work unless you have yourself added as a contact. This will not be the
case for most people. This commit addresses the issue by
short-circuiting the chat message addition and only adds the message to
the overall index for later cache checks and deletions.